Move bughunt report types into module
